MongoDB - Added Querydsl sample.

+++ mongodb/src/main/java/com/oreilly/springdata/mongodb/order/OrderRepository.java
@@ -15,7 +15,6 @@
  */
 package com.oreilly.springdata.mongodb.order;
 
-import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.repository.PagingAndSortingRepository;
@@ -23,9 +22,17 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import com.oreilly.springdata.mongodb.core.Customer;
 
 /**
+ * Repository to access {@link Order}s.
+ * 
  * @author Oliver Gierke
  */
-public interface OrderRepository extends PagingAndSortingRepository<Order, BigInteger> {
+public interface OrderRepository extends PagingAndSortingRepository<Order, Long> {
 
+	/**
+	 * Returns all {@link Order}s of the given {@link Customer}.
+	 * 
+	 * @param customer
+	 * @return
+	 */
 	List<Order> findByCustomer(Customer customer);
 }
